requirements.txt
Cython==3.0.11
fastapi==0.115.4
pydantic==2.7.1
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
uvicorn==0.32.0  
openpyxl
python-dotenv
pymysql
cryptography